Update diun.yml (Diun role)
removed added space from after #s in the diun.yml file under "files" so when you remove comments, the role will not fail. format should be 
```yaml
notif:
  discord:
    webhookURL: https://discord.com/api/webhooks/yes/no
    ...
```
not 
```yaml
 notif
   discord:
     webhookURL:
     ...
```
this could be intentional, so please forgive if this is redundant. Just made starting the role take a while, I had to mess with it more than I'd have needed to otherwise. : )