Nzbget: Fix unpause task creation
Unpause task was failing to be created for new installs. Settings change was successful as part of the role, however NZBGet will not load Task2 if Task1 is not present.
Attempting to create Task1 to save and reload removes the Task2 entry from the config entirely.

Changed this replacement back to Task1, in line with previous CB Master behaviour, unclear if there was any particular reason this was changed to Task2 on CB Develop as no other tasks were created.